Convert to a caugi_graph

Convert an object to a caugi_graph. The object can be a graphNEL, “matrix, or a data frame with from`, `to`, and `edge_type` columns.

Usage

as_caugi(
  x,
  class = c("DAG", "PDAG", "PAG", "Unknown"),
  simple = TRUE,
  build = TRUE,
  collapse = FALSE,
  collapse_to = "---",
  ...
)

Arguments

x

An object to convert to a caugi_graph.

class

"DAG", "PDAG", "PAG", or "Unknown". "PAG" is only supported for integer coded matrices.

simple

logical. If TRUE (default) the graph will be simple (no multiple edges or self-loops).

build

logical. If TRUE (default) build the graph now, otherwise build lazily on first query or when using build().

collapse

logical. If TRUE collapse mutual directed edges to undirected edges. Default is FALSE.

collapse_to

Character string to use as the edge glyph when collapsing. Should be a registered symmetrical edge glyph. Default is "---".

...

Additional arguments passed to specific methods.

Value

A caugi_graph object.

Details

For matrices, as_caugi assumes that the rows are the from nodes and the columns are the to nodes. Thus, for a graph, G: A –> B, we would have that G["A", "B"] == 1 and G["B", "A"] == 0. For PAGs, the integer codes are as follows (as used in pcalg):

0: no edge

1: tail (e.g., A o-- B or A --- B)

2: arrowhead (e.g., A --> B or A o-> B)

3: circle (e.g., A o-o B)