Create a caugi_graph from edge expressions.

Create a caugi_graph from a series of edge expressions using infix operators. Nodes can be specified as symbols, strings, or numbers.

The following edge operators are supported by default:

%-->% for directed edges (A –> B)

%---% for undirected edges (A — B)

%<->% for bidirected edges (A <-> B)

%o->% for partially directed edges (A o-> B)

%o--% for partially undirected edges (A o– B)

%o-o% for partial edges (A o-o B)

You can register additional edge types using register_caugi_edge().

Usage

caugi_graph(
  ...,
  simple = TRUE,
  build = TRUE,
  class = c("Unknown", "DAG", "PDAG")
)

Arguments

...

Edge expressions using the supported infix operators, or nodes given by symbols or strings. Multiple edges can be combined using +: A --> B + C, indicating an edge from A to both B and C. Nodes can also be grouped using c(...) or parentheses.

simple

Logical; if TRUE (default), the graph is a simple graph, and the function will throw an error if the input contains parallel edges or self-loops.

build

Logical; if TRUE (default), the graph will be built using the Rust backend. If FALSE, the graph will not be built, and the Rust backend cannot be used. The graph will build, when queries are made to the graph or if calling build(). Note: Even if build = TRUE, if no edges or nodes are provided, the graph will not be built and the pointer will be NULL.

class

Character; one of "Unknown", "DAG", or "PDAG".

Value

A caugi_graph S7 object containing the nodes, edges, and a pointer to the underlying Rust graph structure.